♻️ Fold plan actions into their plan cards
The subscription actions, upgrade button, and period end label were
single-use client leaves; inlining them makes each plan card fully
self-contained at the cost of the cards becoming client components
(HTML still server-rendered, data still fetched in the page).
